Move common parts of the data models into superclasses to prevent duplication of code.
