+++ discograph/app.py
@@ -1,4 +1,3 @@
-#! /usr/bin/env python
 import atexit
 import os
 import traceback
