+++ tests/integration/library/sqlite/test_sqlite_entity_structural_roles_to_relations.py
