+++ tests/integration/test_ui.py
@@ -1,5 +1,5 @@
 from discograph.app import app
-from tests.app_test_case import AppTestCase
+from tests.integration.app_test_case import AppTestCase
 
 
 class TestUI(AppTestCase):
