+++ tests/unit/library/test_credit_role_from_element.py
